fix table student_group, teacher_group, user_skill and fix method updateUser

+++ src/main/java/ova/example/adminpanel/models/AttestationTheme.java
@@ -17,7 +17,7 @@ public class AttestationTheme {
     private long id;
 
     @Column(name = "course_id")
-    private int courseId;
+    private Long courseId;
 
     private String theme;
 }
